Fixed local repo
So I made a local repo, and it didn't like me when I made a README on the remote. I've now had to exclude the NodeJS modules folder to avoid using Git LFS. I don't mind using it, but, you'll need to install NodeJS anyway if you intend to work on this application. This certainly isn't the worst way to make sure you initialise the project properly on your system.